fix(scripts): Add dynamic import for PQueue to fix module error
